Refactor the data fetching, add comments
